T1: weight int -> double

+++ Module_4/T1-mammals/src/cat.cpp
@@ -1,5 +1,6 @@
 #include "cat.hpp"
 
-Cat::Cat(const std::string &name, int weight) : Mammal(weight), name_(name) {}
+Cat::Cat(const std::string &name, double weight)
+    : Mammal(weight), name_(name) {}
 const std::string &Cat::GetName() const { return name_; }
 std::string Cat::MakeSound() { return "Meow"; }
